+++ hookee/cli.py
@@ -3,13 +3,13 @@ import platform
 import click
 
 from hookee import pluginmanager
-from hookee.climanager import CliManager
+from hookee.hookeemanager import HookeeManager
 
 from future.utils import iteritems
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "1.0.1"
+__version__ = "1.1.0"
 
 
 @click.group(invoke_without_command=True)
@@ -56,12 +56,13 @@ def hookee(ctx, **kwargs):
     if kwargs.get("subdomain") and kwargs.get("hostname"):
         ctx.fail("Can't give both --subdomain and --hostname.")
 
-    cli_manager = CliManager(ctx, load_plugins=ctx.invoked_subcommand not in ["enable-plugin", "disable-plugin",
-                                                                              "available-plugins", "enabled-plugins"])
-    ctx.obj["cli_manager"] = cli_manager
+    hookee_manager = HookeeManager(ctx, load_plugins=ctx.invoked_subcommand not in ["enable-plugin", "disable-plugin",
+                                                                                    "available-plugins",
+                                                                                    "enabled-plugins"])
+    ctx.obj["hookee_manager"] = hookee_manager
 
     if ctx.invoked_subcommand is None:
-        cli_manager.start()
+        hookee_manager.run()
 
 
 @hookee.command()
@@ -70,9 +71,9 @@ def start(ctx):
     """
     Start hookee.
     """
-    cli_manager = ctx.obj["cli_manager"]
+    hookee_manager = ctx.obj["hookee_manager"]
 
-    cli_manager.start()
+    hookee_manager.run()
 
 
 @hookee.command(
@@ -86,7 +87,7 @@ def update_config(ctx, key, value):
     Update the default value for a config. Any passable arg to hookee can also be given here to set its
     default in the config so it doesn't need to be passed to the hookee each time.
     """
-    cli_manager = ctx.obj["cli_manager"]
+    hookee_manager = ctx.obj["hookee_manager"]
 
     if key == "plugins":
         ctx.fail("Enable and disable plugins through the `enable-plugin` and `disable-plugin` commands.")
@@ -97,9 +98,9 @@ def update_config(ctx, key, value):
         value = int(value)
 
     try:
-        cli_manager.config.set(key, value)
+        hookee_manager.config.set(key, value)
 
-        cli_manager.print_util.print_config_update(
+        hookee_manager.print_util.print_config_update(
             "The default value for \"{}\" has been updated in the config.".format(key))
     except KeyError:
         ctx.fail("No such key exists in the config: {}".format(key))
@@ -112,13 +113,13 @@ def enable_plugin(ctx, plugin):
     """
     Enable the given plugin by default.
     """
-    cli_manager = ctx.obj["cli_manager"]
+    hookee_manager = ctx.obj["hookee_manager"]
 
-    cli_manager.plugin_manager.get_plugin(plugin)
+    hookee_manager.plugin_manager.get_plugin(plugin)
 
-    cli_manager.config.append("plugins", plugin)
+    hookee_manager.config.append("plugins", plugin)
 
-    cli_manager.print_util.print_config_update("Plugin \"{}\" has been enabled.".format(plugin))
+    hookee_manager.print_util.print_config_update("Plugin \"{}\" has been enabled.".format(plugin))
 
 
 @hookee.command()
@@ -128,14 +129,14 @@ def disable_plugin(ctx, plugin):
     """
     Disable the given plugin by default.
     """
-    cli_manager = ctx.obj["cli_manager"]
+    hookee_manager = ctx.obj["hookee_manager"]
 
     if plugin in pluginmanager.REQUIRED_PLUGINS:
         ctx.fail("Can't disable the plugin \"{}\".".format(plugin))
 
-    cli_manager.config.remove("plugins", plugin)
+    hookee_manager.config.remove("plugins", plugin)
 
-    cli_manager.print_util.print_config_update("Plugin \"{}\" is disabled.".format(plugin))
+    hookee_manager.print_util.print_config_update("Plugin \"{}\" is disabled.".format(plugin))
 
 
 @hookee.command()
@@ -144,12 +145,11 @@ def available_plugins(ctx):
     """
     List all available plugins.
     """
-    cli_manager = ctx.obj["cli_manager"]
+    hookee_manager = ctx.obj["hookee_manager"]
 
-    plugins = cli_manager.plugin_manager.available_plugins()
+    plugins = hookee_manager.plugin_manager.available_plugins()
 
-    click.secho("\nAvailable Plugins: {}\n".format(plugins),
-                fg="green")
+    hookee_manager.print_util.print_basic("\nAvailable Plugins: {}\n".format(plugins), fg="green")
 
 
 @hookee.command()
@@ -158,11 +158,11 @@ def enabled_plugins(ctx):
     """
     List all enabled plugins.
     """
-    cli_manager = ctx.obj["cli_manager"]
+    hookee_manager = ctx.obj["hookee_manager"]
 
-    plugins = cli_manager.plugin_manager.enabled_plugins()
+    plugins = hookee_manager.plugin_manager.enabled_plugins()
 
-    click.secho("\nEnabled Plugins: {}\n".format(plugins), fg="green")
+    hookee_manager.print_util.print_basic("\nEnabled Plugins: {}\n".format(plugins), fg="green")
 
 
 if __name__ == "__main__":