+++ hookee/plugins/request_headers.py
@@ -1,24 +1,22 @@
-import click
-
 from hookee import pluginmanager
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.7"
+__version__ = "1.1.0"
 
 plugin_type = pluginmanager.REQUEST_PLUGIN
 print_util = None
 
 
-def setup(cli_manager):
+def setup(hookee_manager):
     global print_util
 
-    print_util = cli_manager.print_util
+    print_util = hookee_manager.print_util
 
 
 def run(request):
     if request.headers and "X-Forwarded-For" in request.headers:
-        click.secho("Client IP: {}".format(request.headers.get("X-Forwarded-For")), fg="magenta")
+        print_util.print_basic("Client IP: {}".format(request.headers.get("X-Forwarded-For")), fg="magenta")
     if request.headers:
         print_util.print_dict("Headers", dict(request.headers), fg="magenta")