+++ tests/test_hookee_manager.py
@@ -5,10 +5,10 @@ from tests.managedtestcase import ManagedTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.12"
+__version__ = "1.1.0"
 
 
-class TestCliManager(ManagedTestCase):
+class TestHookeeManager(ManagedTestCase):
     def test_http_get_query_params(self):
         # GIVEN
         params = {"param_1": ["param_value_1"]}